Seperating the logic a bit
